+++ demo/ProgressBarDemo.qml
@@ -1,31 +1,46 @@
 import QtQuick 2.0
+import QtQuick.Layouts 1.1
 import Material 0.1
 
 Item {
-    Column {
+    implicitHeight: grid.implicitHeight + units.dp(40)
+    GridLayout {
+        id: grid
         anchors.centerIn: parent
-        spacing: units.dp(25)
+        columns: 2
+        columnSpacing: units.dp(20)
+        rowSpacing: units.dp(20)
 
         Label {
             text: "Determinate"
         }
 
         ProgressBar {
-            id: progressBar1
-            width: units.dp(50)
+            Layout.fillWidth: true
+            color: theme.accentColor
 
-            SequentialAnimation on progress {
+            SequentialAnimation on value {
                 running: true
                 loops: NumberAnimation.Infinite
-
-                PauseAnimation { duration: 1000 } // This puts a bit of time between the loop
 
                 NumberAnimation {
                     duration: 3000
                     from: 0
                     to: 1
                 }
+
+                PauseAnimation { duration: 1000 } // This puts a bit of time between the loop
             }
+        }
+
+        Label {
+            text: "Indeterminate"
+        }
+
+        ProgressBar {
+            Layout.fillWidth: true
+            color: theme.accentColor
+            indeterminate: true
         }
     }
 }